Better handling of redirects by 'header()'
