+++ src/Controller/ProceduralRedirect.php
@@ -1,0 +1,6 @@
+<?php
+
+echo "Some text";
+
+header("Location: /DirectAccess.php");
+exit;
